+++ src/components/pages/Dashboard/User/MyOrders/MyOrders.js
@@ -1,13 +1,11 @@
 import { signOut } from 'firebase/auth';
 import React, { useEffect, useState } from 'react';
-import { Button, Modal } from 'react-bootstrap';
 import { useAuthState } from 'react-firebase-hooks/auth';
 import { Link, useNavigate } from 'react-router-dom';
 import { auth } from '../../../../../firebase.init';
 
 const MyOrders = () => {
     const [orders, setOrders] = useState([]);
-    const [show, setShow] = useState(false);
     const [user] = useAuthState(auth);
     const navigate = useNavigate()
 
@@ -35,24 +33,23 @@ const MyOrders = () => {
         }
     }, [user]);
 
-    const handleClose = () => setShow(false);
-    const handleShow = () => setShow(true);
+    const handleDeleteProduct = (id) => {
+        const confirmToDelete = window.confirm('Are you sure, want to delete this product?');
+        if (confirmToDelete) {
+            const url = `http://localhost:5000/order/${id}`;
 
-    const handleConfirm = (id) => {
-        setShow(false);
-        const url = `http://localhost:5000/order/${id}`;
             fetch(url, {
                 method: 'DELETE'
             })
-            .then(res => res.json())
-            .then(data =>{
-                if(data.deletedCount > 0){
-                    
-                    const remaining = orders.filter(order => order._id !== id);
-                    setOrders(remaining);
-                }
-            })
+                .then(res => res.json())
+                .then(data => {
+                    if (data.deletedCount > 0) {
+                        const remainingOrders = orders.filter(order => order._id !== id);
+                        setOrders(remainingOrders);
+                    }
 
+                })
+        }
     }
 
 
@@ -97,24 +94,11 @@ const MyOrders = () => {
                                                     !order.paid && <Link to={`/dashboard/payment/${order._id}`} ><span className=" btn bg-btn text-white p-1">pay</span></Link>
                                                 }
                                                 {
-                                                    order.paid && <span className="bg-success text-white p-1">paid</span>
+                                                    order.paid && <span className="bg-success text-white p-1">Pending</span>
                                                 }
-                                                <span onClick={handleShow} className="btn btn-danger text-white ms-1 p-1">cancel</span></td>
+                                                <span onClick={() => handleDeleteProduct(order?._id)} className="btn btn-danger text-white ms-1 p-1">cancel</span></td>
 
-                                            <Modal show={show} onHide={handleClose}>
-                                                <Modal.Header closeButton>
-                                                    {/* <Modal.Title></Modal.Title> */}
-                                                </Modal.Header>
-                                                <Modal.Body>Are you sure, want to cancel this order?</Modal.Body>
-                                                <Modal.Footer>
-                                                    <Button variant="secondary" onClick={handleClose}>
-                                                        No
-                                                    </Button>
-                                                    <Button variant="primary" onClick={() => handleConfirm(order._id)}>
-                                                        Yes
-                                                    </Button>
-                                                </Modal.Footer>
-                                            </Modal>
+                                            
 
                                         </tr>
                                     )
